+++ BitmovinAnalyticsCollector/Classes/Collector/stateMachines/StateMachine.swift
@@ -74,6 +74,7 @@ public class StateMachine {
         let performTransition = checkUnallowedTransitions(destinationState: destinationState)
 
         if performTransition {
+            print("[StateMachine] Transitioning from state \(state) to \(destinationState)")
             let timestamp = Date().timeIntervalSince1970Millis
             let previousState = state
             videoTimeEnd = time
@@ -137,7 +138,8 @@ public class StateMachine {
         }
         clearVideoStartFailedTimer()
         
-        videoStartFailedWorkItem = DispatchWorkItem {
+        videoStartFailedWorkItem = DispatchWorkItem { [weak self] in
+            guard let self = self else { return }
             self.errorData = ErrorData.ANALYTICS_VIDEOSTART_TIMEOUT_REACHED
             self.onPlayAttemptFailed(withReason: VideoStartFailedReason.timeout)
         }