+++ BitmovinAnalyticsCollector/Classes/Collector/stateMachines/util/QualityChangeCounter.swift
@@ -4,17 +4,18 @@ public class QualityChangeCounter {
     private static var kAnalyticsQualityChangeThreshold = 50
     private static var kAnalyticsQualityChangeResetIntervalSeconds: TimeInterval = 60 * 60
     private static var kAnalyticsQualityChangeIntervalId = "com.bitmovin.analytics.core.utils.QualityChangeCounter"
-     
+    private var queue = DispatchQueue.init(label: QualityChangeCounter.kAnalyticsQualityChangeIntervalId)
+    
     private var qualityResetWorkItem: DispatchWorkItem?
     private var qualityChangeCounter = 0
     
     func startInterval() {
         resetInterval()
-        qualityResetWorkItem = DispatchWorkItem {
+        qualityResetWorkItem = DispatchWorkItem { [weak self] in
+            guard let self = self else { return }
             self.qualityChangeCounter = 0
         }
-        
-        DispatchQueue.init(label: QualityChangeCounter.kAnalyticsQualityChangeIntervalId).asyncAfter(deadline: .now() + QualityChangeCounter.kAnalyticsQualityChangeResetIntervalSeconds, execute: qualityResetWorkItem!)
+        queue.asyncAfter(deadline: .now() + QualityChangeCounter.kAnalyticsQualityChangeResetIntervalSeconds, execute: qualityResetWorkItem!)
     }
     
     func resetInterval() {